+++ src/lib/Prefs/PrefsKeys.js
@@ -173,6 +173,18 @@ var PrefsKeys = class
 
         this.setKey(
             'visibility',
+            'screen-recording-indicator',
+            'GtkSwitch',
+            this._shellVersion >= 43,
+            {
+                default: true,
+                minimal: true,
+                superminimal: true,
+            }
+        );
+
+        this.setKey(
+            'visibility',
             'search',
             'GtkSwitch',
             true,